move client only types to client file
