+++ src/internal/type.ts
@@ -57,14 +57,6 @@ export interface IRequest {
 
 export type ICanonicalRequest = string
 
-export type NoResultCallback = (error: unknown) => void
-
-export interface RemoveOptions {
-  versionId?: string
-  governanceBypass?: boolean
-  forceDelete?: boolean
-}
-
 export interface IncompleteUploadedBucketItem {
   key: string
   uploadId: string
